[WIP] "Adventure Demo" level and guide
The Adventure Demo provides design patterns and an example of how to use Ink and Inkpot to manage narrative text for a Narrative Adventure game in Unreal Engine.

It covers design patterns for implementing:
- Dialogue
- Player Journal
- Quests

Here's a summary of changes:

## AdventureDemo.ink
The ink file for the Adventure Demo.

## AdventureDemo.umap and .uasset
The actual Adventure Demo level.

There's some basic path-switching logic in the level BP.

## WBP_Display_Adventure

An expanded version of WBP_Display that includes:
- The Journal panel, Journal button, and JournalUpdated notification
- Logic for opening/closing the journal
- Logic for showing Journal Updates, using a Story Change Delegate
- Visual Updates to the main story box, and logic to hide it when there's no narrative text to show
- Logic to customize the timer for each line using a `linetime: ` tag in Ink

A lot of this code is duplicated from Inkpot Demo. This is inelegant, but @TCR-Nick  is about to overhaul the demo level, so it's easier to keep the Adventure Demo separate until his changes are done.

## BP_AdventureDemoGameMode:

This is a copy of `InkpotDemoGameMode` that points to `WBP_Display_Adventure` in place of `WBP_Display` widget blueprint.

Again, inelegant--but it's better to wait until the main Demo level is updated before consolidating.

## README-AdventureDemo.md

Honestly a bit much. Documentation escaped containment.